requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
eventlet==0.35.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.2
